fix(hide-activities): get consistent behavior in the first workspace
If the activities view was activated by the extension, it gets disabled
once the user goes back to a workspace that is not empty. However, when
no windows are open, the user may go back to the first workspace still
empty. Now, the activities view also gets disabled in this case.