fix(show-activities): fix activities not showing on certain multimonitor setups
If GNOME is configured to only have workspaces on the primary display
(while other displays have a static workspace for themselves), only
accept windows in the primary monitor. Without this, the activities view
would only be shown when all windows in all monitors were closed, which
is not the desired behavior.