fix(temporary-windows): ignore closing of temporary windows
Temporary windows are windows that are opened, then automatically closed
by their own programs. Usually, they are shown before an application
fully opens, as is the case with DBeaver and LibreOffice. Steam also has
tons of temporary windows: for updating itself, for signing in, for
launching games, etc.

The problem is that these temporary windows disrupt Foresight's
detection of empty workspaces. If they are opened on an empty workspace,
that workspace soon becomes empty again (so Foresight opens the
Activities view), and then not empty when the main window of the
application opens. So users have to manually leave the Activities view.

To address this issue, Foresight will attempt to detect temporary
windows using patterns in their properties (title, class, etc). Since it
can't possible match all temporary windows in existence, this is only a
partial fix for now.

Related: https://github.com/pesader/gnome-shell-extension-foresight/issues/9